backend/requirements.txt
google-adk==1.2.1
fastapi==0.115.0
uvicorn[standard]>=0.34.0
python-dotenv==1.0.1
websockets==13.1
python-multipart==0.0.12
certifi==2024.8.30
google-cloud-aiplatform>=1.38.0 